refactoring dialogs
