+++ src/Components/Dialogs/GetPerson/getMessage.module.css
@@ -1,0 +1,3 @@
+.item{
+    margin: 10px 10px;
+}
